requirements.txt
requests==2.32.3
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.6.0.zip
tenacity==9.0.0